Add files via upload
Objective
The objective of this final project is to design and implement a web application using Docker that 
includes a web server, a database, and the ability to display team data on the website. Students will gain 
hands-on experience in containerization, communication between containers using Docker Compose, 
and web development.